cras_cpp_common: Completely reworked log_utils to use macros instead of functions.
This was needed because of the static log_location variables inside ROS_ macros - e.g. _ONCE was only triggered once regardless of where was it called from. There were also not so helpful file:line data in the logged messages.

Backwards compatibility was kept 99%, but there are subtle cases where it will fail - e.g. if there was `this->log->logError()` right after an `if` or `else` without braces.